+++ dplanner/lib/services/club_api_service.dart
@@ -65,37 +65,6 @@ class ClubApiService {
     throw ErrorDescription(errorMessage);
   }
 
-  /// POST: /clubs/(_.club_id)/join [클럽 가입하기] 클럽 멤버 가입하기
-  static Future<ClubMemberModel> postClubMember(
-      {required int clubId, required String name, required String info}) async {
-    final url = Uri.parse('$baseUrl/clubs/$clubId/join');
-    const storage = FlutterSecureStorage();
-
-    String? accessToken = await storage.read(key: accessTokenKey);
-
-    final response = await http.post(
-      url,
-      headers: <String, String>{
-        'Content-Type': 'application/json',
-        'Authorization': 'Bearer $accessToken',
-      },
-      body: jsonEncode({
-        "name": name,
-        "info": info,
-      }),
-    );
-
-    if (response.statusCode == 201) {
-      return ClubMemberModel.fromJson(
-          jsonDecode(utf8.decode(response.bodyBytes))['data']);
-    }
-
-    // 예외 처리; 메시지를 포함한 예외를 던짐
-    String errorMessage = jsonDecode(response.body)['message'] ?? 'Error';
-    print(errorMessage);
-    throw ErrorDescription(errorMessage);
-  }
-
   /// GET: /clubs?memberId=(_.member_id) [내 클럽 목록] 내 클럽 목록 불러오기
   static Future<List<ClubModel>> getClubList() async {
     const storage = FlutterSecureStorage();
